variables and constants get / set functions, and more bind code cleanup
