+++ bind/package.go
@@ -329,237 +329,3 @@ func (p *Package) Lookup(o types.Object) (Object, bool) {
 	obj, ok := p.objs[o.Name()]
 	return obj, ok
 }
-
-// Protocol encodes the various protocols a python type may implement
-type Protocol int
-
-const (
-	ProtoStringer Protocol = 1 << iota
-)
-
-// Struct collects informations about a go struct.
-type Struct struct {
-	pkg *Package
-	sym *symbol
-	obj *types.TypeName
-
-	id    string
-	doc   string
-	ctors []Func
-	meths []Func
-
-	prots Protocol
-}
-
-func newStruct(p *Package, obj *types.TypeName) (Struct, error) {
-	sym := p.syms.symtype(obj.Type())
-	if sym == nil {
-		panic(fmt.Errorf("no such object [%s] in symbols table", obj.Id()))
-	}
-	sym.doc = p.getDoc("", obj)
-	s := Struct{
-		pkg: p,
-		sym: sym,
-		obj: obj,
-	}
-	return s, nil
-}
-
-func (s Struct) Package() *Package {
-	return s.pkg
-}
-
-func (s Struct) ID() string {
-	return s.sym.id
-}
-
-func (s Struct) Doc() string {
-	return s.sym.doc
-}
-
-func (s Struct) GoType() types.Type {
-	return s.sym.GoType()
-}
-
-func (s Struct) GoName() string {
-	return s.sym.goname
-}
-
-func (s Struct) Struct() *types.Struct {
-	return s.sym.GoType().Underlying().(*types.Struct)
-}
-
-// A Signature represents a (non-builtin) function or method type.
-type Signature struct {
-	ret  []*Var
-	args []*Var
-	recv *Var
-}
-
-func newSignatureFrom(pkg *Package, sig *types.Signature) *Signature {
-	var recv *Var
-	if sig.Recv() != nil {
-		recv = newVarFrom(pkg, sig.Recv())
-	}
-
-	return &Signature{
-		ret:  newVarsFrom(pkg, sig.Results()),
-		args: newVarsFrom(pkg, sig.Params()),
-		recv: recv,
-	}
-}
-
-func newSignature(pkg *Package, recv *Var, params, results []*Var) *Signature {
-	return &Signature{
-		ret:  results,
-		args: params,
-		recv: recv,
-	}
-}
-
-func (sig *Signature) Results() []*Var {
-	return sig.ret
-}
-
-func (sig *Signature) Params() []*Var {
-	return sig.args
-}
-
-func (sig *Signature) Recv() *Var {
-	return sig.recv
-}
-
-// Func collects informations about a go func/method.
-type Func struct {
-	pkg  *Package
-	sig  *Signature
-	typ  types.Type
-	name string
-
-	id   string
-	doc  string
-	ret  types.Type // return type, if any
-	err  bool       // true if original go func has comma-error
-	ctor bool       // true if this is a newXXX function
-}
-
-func newFuncFrom(p *Package, parent string, obj types.Object, sig *types.Signature) (Func, error) {
-	haserr := false
-	res := sig.Results()
-	var ret types.Type
-
-	switch res.Len() {
-	case 2:
-		if !isErrorType(res.At(1).Type()) {
-			return Func{}, fmt.Errorf(
-				"bind: second result value must be of type error: %s",
-				obj,
-			)
-		}
-		haserr = true
-		ret = res.At(0).Type()
-
-	case 1:
-		if isErrorType(res.At(0).Type()) {
-			haserr = true
-			ret = nil
-		} else {
-			ret = res.At(0).Type()
-		}
-	case 0:
-		ret = nil
-	default:
-		return Func{}, fmt.Errorf("bind: too many results to return: %v", obj)
-	}
-
-	id := obj.Pkg().Name() + "_" + obj.Name()
-	if parent != "" {
-		id = obj.Pkg().Name() + "_" + parent + "_" + obj.Name()
-	}
-
-	return Func{
-		pkg:  p,
-		sig:  newSignatureFrom(p, sig),
-		typ:  obj.Type(),
-		name: obj.Name(),
-		id:   id,
-		doc:  p.getDoc(parent, obj),
-		ret:  ret,
-		err:  haserr,
-	}, nil
-}
-
-func (f Func) Package() *Package {
-	return f.pkg
-}
-
-func (f Func) ID() string {
-	return f.id
-}
-
-func (f Func) Doc() string {
-	return f.doc
-}
-
-func (f Func) GoType() types.Type {
-	return f.typ
-}
-
-func (f Func) GoName() string {
-	return f.name
-}
-
-func (f Func) GoFmt() string {
-	return f.pkg.Name() + "." + f.name
-}
-
-func (f Func) Signature() *Signature {
-	return f.sig
-}
-
-func (f Func) Return() types.Type {
-	return f.ret
-}
-
-type Const struct {
-	pkg *Package
-	sym *symbol
-	obj *types.Const
-	id  string
-	doc string
-	f   Func
-}
-
-func newConst(p *Package, o *types.Const) Const {
-	pkg := o.Pkg()
-	sym := p.syms.symtype(o.Type())
-	id := pkg.Name() + "_" + o.Name()
-	doc := p.getDoc("", o)
-
-	res := []*Var{newVar(p, o.Type(), "ret", o.Name(), doc)}
-	sig := newSignature(p, nil, nil, res)
-	fct := Func{
-		pkg:  p,
-		sig:  sig,
-		typ:  nil,
-		name: o.Name(),
-		id:   id + "_get",
-		doc:  doc,
-		ret:  o.Type(),
-		err:  false,
-	}
-
-	return Const{
-		pkg: p,
-		sym: sym,
-		obj: o,
-		id:  id,
-		doc: doc,
-		f:   fct,
-	}
-}
-
-func (c Const) ID() string         { return c.id }
-func (c Const) Doc() string        { return c.doc }
-func (c Const) GoName() string     { return c.obj.Name() }
-func (c Const) GoType() types.Type { return c.obj.Type() }
