+++ bind/symtab.go
@@ -695,11 +695,11 @@ func init() {
 			gotyp:   look("bool").Type(),
 			kind:    skType | skBasic,
 			goname:  "bool",
-			cgoname: "bool",
-			cpyname: "bool",
+			cgoname: "C.char",
+			cpyname: "uint8_t",
 			pysig:   "bool",
-			go2py:   "cgopy_cnv_go2py_bool",
-			py2go:   "cgopy_cnv_py2go_bool",
+			go2py:   "boolGoToPy",
+			py2go:   "boolPyToGo",
 			zval:    "false",
 		},
