+++ gen/test.py
@@ -9,23 +9,23 @@ import hi
 print("--- doc(hi)...")
 print(hi.__doc__)
 
-#print("--- hi.GetUniverse():", hi.GetUniverse())
-#print("--- hi.GetVersion():", hi.GetVersion())
+print("--- hi.GetUniverse():", hi.GetUniverse())
+print("--- hi.GetVersion():", hi.GetVersion())
 
-#print("--- hi.GetDebug():",hi.GetDebug())
-#print("--- hi.SetDebug(true)")
-#hi.SetDebug(True)
-#print("--- hi.GetDebug():",hi.GetDebug())
-#print("--- hi.SetDebug(false)")
-#hi.SetDebug(False)
-#print("--- hi.GetDebug():",hi.GetDebug())
+print("--- hi.GetDebug():",hi.GetDebug())
+print("--- hi.SetDebug(true)")
+hi.SetDebug(True)
+print("--- hi.GetDebug():",hi.GetDebug())
+print("--- hi.SetDebug(false)")
+hi.SetDebug(False)
+print("--- hi.GetDebug():",hi.GetDebug())
 
-#print("--- hi.GetAnon():",hi.GetAnon())
-#anon = hi.NewPerson('you',24)
-#print("--- new anon:",anon)
-#print("--- hi.SetAnon(hi.NewPerson('you', 24))...")
-#hi.SetAnon(anon)
-#print("--- hi.GetAnon():",hi.GetAnon())
+print("--- hi.GetAnon():",hi.GetAnon())
+anon = hi.NewPerson('you',24)
+print("--- new anon:",anon)
+print("--- hi.SetAnon(hi.NewPerson('you', 24))...")
+hi.SetAnon(anon)
+print("--- hi.GetAnon():",hi.GetAnon())
 
 print("--- doc(hi.Hi)...")
 print(hi.Hi.__doc__)
